Se agrega el obtener 5 tareas aleatorias desde API, agregando provideHttpClient en appconfig para habilitar el HttpClient y poder consultar los datos
