+++ .angular/cache/19.2.15/faser-prueba-tecnica-v2/vite/deps/@angular_common_http.js
@@ -1,0 +1,85 @@
+import {
+  FetchBackend,
+  HTTP_INTERCEPTORS,
+  HTTP_ROOT_INTERCEPTOR_FNS,
+  HTTP_TRANSFER_CACHE_ORIGIN_MAP,
+  HttpBackend,
+  HttpClient,
+  HttpClientJsonpModule,
+  HttpClientModule,
+  HttpClientXsrfModule,
+  HttpContext,
+  HttpContextToken,
+  HttpErrorResponse,
+  HttpEventType,
+  HttpFeatureKind,
+  HttpHandler,
+  HttpHeaderResponse,
+  HttpHeaders,
+  HttpInterceptorHandler,
+  HttpParams,
+  HttpRequest,
+  HttpResponse,
+  HttpResponseBase,
+  HttpStatusCode,
+  HttpUrlEncodingCodec,
+  HttpXhrBackend,
+  HttpXsrfTokenExtractor,
+  JsonpClientBackend,
+  JsonpInterceptor,
+  REQUESTS_CONTRIBUTE_TO_STABILITY,
+  httpResource,
+  provideHttpClient,
+  withFetch,
+  withHttpTransferCache,
+  withInterceptors,
+  withInterceptorsFromDi,
+  withJsonpSupport,
+  withNoXsrfProtection,
+  withRequestsMadeViaParent,
+  withXsrfConfiguration
+} from "./chunk-2LFFPLM5.js";
+import "./chunk-FRXTEGWH.js";
+import "./chunk-LKBYGJ3F.js";
+export {
+  FetchBackend,
+  HTTP_INTERCEPTORS,
+  HTTP_TRANSFER_CACHE_ORIGIN_MAP,
+  HttpBackend,
+  HttpClient,
+  HttpClientJsonpModule,
+  HttpClientModule,
+  HttpClientXsrfModule,
+  HttpContext,
+  HttpContextToken,
+  HttpErrorResponse,
+  HttpEventType,
+  HttpFeatureKind,
+  HttpHandler,
+  HttpHeaderResponse,
+  HttpHeaders,
+  HttpParams,
+  HttpRequest,
+  HttpResponse,
+  HttpResponseBase,
+  HttpStatusCode,
+  HttpUrlEncodingCodec,
+  HttpXhrBackend,
+  HttpXsrfTokenExtractor,
+  JsonpClientBackend,
+  JsonpInterceptor,
+  httpResource,
+  provideHttpClient,
+  withFetch,
+  withInterceptors,
+  withInterceptorsFromDi,
+  withJsonpSupport,
+  withNoXsrfProtection,
+  withRequestsMadeViaParent,
+  withXsrfConfiguration,
+  HTTP_ROOT_INTERCEPTOR_FNS as ɵHTTP_ROOT_INTERCEPTOR_FNS,
+  HttpInterceptorHandler as ɵHttpInterceptingHandler,
+  HttpInterceptorHandler as ɵHttpInterceptorHandler,
+  REQUESTS_CONTRIBUTE_TO_STABILITY as ɵREQUESTS_CONTRIBUTE_TO_STABILITY,
+  withHttpTransferCache as ɵwithHttpTransferCache
+};
