+++ .angular/cache/19.2.15/faser-prueba-tecnica-v2/vite/deps/@angular_platform-browser.js
@@ -34,10 +34,12 @@ import {
   withI18nSupport,
   withIncrementalHydration,
   withNoHttpTransferCache
-} from "./chunk-5OPZPMAB.js";
+} from "./chunk-E2X7F25W.js";
+import "./chunk-2LFFPLM5.js";
 import {
   getDOM
-} from "./chunk-I3C46P22.js";
+} from "./chunk-KOUZRL47.js";
+import "./chunk-FRXTEGWH.js";
 import "./chunk-LKBYGJ3F.js";
 export {
   BrowserModule,
