+++ .angular/cache/19.2.15/faser-prueba-tecnica-v2/vite/deps/@angular_router.js
@@ -1,15 +1,18 @@
 import {
   Title
-} from "./chunk-5OPZPMAB.js";
+} from "./chunk-E2X7F25W.js";
+import "./chunk-2LFFPLM5.js";
 import {
-  DOCUMENT,
   HashLocationStrategy,
   LOCATION_INITIALIZED,
   Location,
   LocationStrategy,
   PathLocationStrategy,
   ViewportScroller
-} from "./chunk-I3C46P22.js";
+} from "./chunk-KOUZRL47.js";
+import {
+  DOCUMENT
+} from "./chunk-FRXTEGWH.js";
 import {
   APP_BOOTSTRAP_LISTENER,
   ApplicationRef,
