+++ .angular/cache/19.2.15/faser-prueba-tecnica-v2/vite/deps_ssr/@angular_platform-browser.js
@@ -35,10 +35,12 @@ import {
   withI18nSupport,
   withIncrementalHydration,
   withNoHttpTransferCache
-} from "./chunk-4KJ25FNC.js";
+} from "./chunk-SY4YM4RU.js";
+import "./chunk-XNYIGXTF.js";
 import {
   getDOM
-} from "./chunk-YT24NAFA.js";
+} from "./chunk-53KWUNRH.js";
+import "./chunk-DFIAP3GC.js";
 import "./chunk-PXXRCHXC.js";
 import "./chunk-YHCV7DAQ.js";
 export {
