+++ .angular/cache/19.2.15/faser-prueba-tecnica-v2/vite/deps_ssr/@angular_platform-server.js
@@ -15,9 +15,11 @@ import {
   renderApplication,
   renderInternal,
   renderModule
-} from "./chunk-SSGKJJWO.js";
-import "./chunk-4KJ25FNC.js";
-import "./chunk-YT24NAFA.js";
+} from "./chunk-FYXEBGAD.js";
+import "./chunk-SY4YM4RU.js";
+import "./chunk-XNYIGXTF.js";
+import "./chunk-53KWUNRH.js";
+import "./chunk-DFIAP3GC.js";
 import "./chunk-PXXRCHXC.js";
 import "./chunk-YHCV7DAQ.js";
 export {
