+++ .angular/cache/19.2.15/faser-prueba-tecnica-v2/vite/deps_ssr/@angular_router.js
@@ -77,9 +77,11 @@ import {
   withRouterConfig,
   withViewTransitions,
   ɵEmptyOutletComponent
-} from "./chunk-Y5BGAT76.js";
-import "./chunk-4KJ25FNC.js";
-import "./chunk-YT24NAFA.js";
+} from "./chunk-CC4F4OUU.js";
+import "./chunk-SY4YM4RU.js";
+import "./chunk-XNYIGXTF.js";
+import "./chunk-53KWUNRH.js";
+import "./chunk-DFIAP3GC.js";
 import "./chunk-PXXRCHXC.js";
 import "./chunk-YHCV7DAQ.js";
 export {
