+++ .angular/cache/19.2.15/faser-prueba-tecnica-v2/vite/deps_ssr/@angular_ssr.js
@@ -14,11 +14,13 @@ import {
   setAngularAppEngineManifest,
   setAngularAppManifest,
   withAppShell
-} from "./chunk-5VBN6X4N.js";
-import "./chunk-SSGKJJWO.js";
-import "./chunk-Y5BGAT76.js";
-import "./chunk-4KJ25FNC.js";
-import "./chunk-YT24NAFA.js";
+} from "./chunk-RMKP4LFV.js";
+import "./chunk-FYXEBGAD.js";
+import "./chunk-CC4F4OUU.js";
+import "./chunk-SY4YM4RU.js";
+import "./chunk-XNYIGXTF.js";
+import "./chunk-53KWUNRH.js";
+import "./chunk-DFIAP3GC.js";
 import "./chunk-PXXRCHXC.js";
 import "./chunk-YHCV7DAQ.js";
 export {
