+++ .angular/cache/19.2.15/faser-prueba-tecnica-v2/vite/deps_ssr/@angular_ssr_node.js
@@ -2,15 +2,17 @@ import { createRequire } from 'module';const require = createRequire(import.meta
 import {
   AngularAppEngine,
   InlineCriticalCssProcessor
-} from "./chunk-5VBN6X4N.js";
+} from "./chunk-RMKP4LFV.js";
 import {
   SERVER_CONTEXT,
   renderApplication,
   renderModule
-} from "./chunk-SSGKJJWO.js";
-import "./chunk-Y5BGAT76.js";
-import "./chunk-4KJ25FNC.js";
-import "./chunk-YT24NAFA.js";
+} from "./chunk-FYXEBGAD.js";
+import "./chunk-CC4F4OUU.js";
+import "./chunk-SY4YM4RU.js";
+import "./chunk-XNYIGXTF.js";
+import "./chunk-53KWUNRH.js";
+import "./chunk-DFIAP3GC.js";
 import "./chunk-PXXRCHXC.js";
 import {
   __async,
