+++ .angular/cache/19.2.15/faser-prueba-tecnica-v2/vite/deps_ssr/chunk-CC4F4OUU.js
@@ -1,16 +1,18 @@
 import { createRequire } from 'module';const require = createRequire(import.meta.url);
 import {
   Title
-} from "./chunk-4KJ25FNC.js";
+} from "./chunk-SY4YM4RU.js";
 import {
-  DOCUMENT,
   HashLocationStrategy,
   LOCATION_INITIALIZED,
   Location,
   LocationStrategy,
   PathLocationStrategy,
   ViewportScroller
-} from "./chunk-YT24NAFA.js";
+} from "./chunk-53KWUNRH.js";
+import {
+  DOCUMENT
+} from "./chunk-DFIAP3GC.js";
 import {
   APP_BOOTSTRAP_LISTENER,
   ApplicationRef,
@@ -5962,4 +5964,4 @@ export {
    * License: MIT
    *)
 */
-//# sourceMappingURL=chunk-Y5BGAT76.js.map
+//# sourceMappingURL=chunk-CC4F4OUU.js.map
