+++ .angular/cache/19.2.15/faser-prueba-tecnica-v2/vite/deps_ssr/chunk-RMKP4LFV.js
@@ -5,16 +5,16 @@ import {
   platformServer,
   renderApplication,
   renderModule
-} from "./chunk-SSGKJJWO.js";
+} from "./chunk-FYXEBGAD.js";
 import {
   ROUTES,
   Router,
   loadChildren
-} from "./chunk-Y5BGAT76.js";
+} from "./chunk-CC4F4OUU.js";
 import {
   APP_BASE_HREF,
   PlatformLocation
-} from "./chunk-YT24NAFA.js";
+} from "./chunk-53KWUNRH.js";
 import {
   ApplicationRef,
   Compiler,
@@ -11987,4 +11987,4 @@ export {
   AngularAppEngine,
   createRequestHandler
 };
-//# sourceMappingURL=chunk-5VBN6X4N.js.map
+//# sourceMappingURL=chunk-RMKP4LFV.js.map
