+++ src/index.html
@@ -3,7 +3,6 @@
 <head>
   <meta charset="utf-8">
   <title>FaserPruebaTecnicaV2</title>
-  <h1>Control de tareas - Luis Franco</h1>
   <base href="/">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="icon" type="image/x-icon" href="favicon.ico">
